+++ src/RcppCyclopsInterface.cpp
@@ -49,6 +49,10 @@ namespace bsccs {
  	{ModelType::FINE_GRAY, "fgr"},
  };
 
+void RcppCcdInterface::logResultsToFile(const std::string& fileName, bool withASE) {
+    ccd->logResults(fileName.c_str(), withASE);
+}
+
 } // namespace bsccs
 
 // [[Rcpp::export(".cyclopsGetModelTypeNames")]]
@@ -167,16 +171,16 @@ void cyclopsSetCensorWeights(SEXP inRcppCcdInterface,
     interface->getCcd().setCensorWeights(&weights[0]);
 }
 
-// [[Rcpp::export(".cyclopsGetPredictiveLogLikelihood")]]
-double cyclopsGetPredictiveLogLikelihood(SEXP inRcppCcdInterface,
-    NumericVector& weights) {
-    using namespace bsccs;
-    XPtr<RcppCcdInterface> interface(inRcppCcdInterface);
-
-    // return interface->getCcd().getPredictiveLogLikelihood(&weights[0]);
-    Rcpp::stop("No longer implemented");
-    return 0.0;
-}
+// // [[Rcpp::export(".cyclopsGetPredictiveLogLikelihood")]]
+// double cyclopsGetPredictiveLogLikelihood(SEXP inRcppCcdInterface,
+//     NumericVector& weights) {
+//     using namespace bsccs;
+//     XPtr<RcppCcdInterface> interface(inRcppCcdInterface);
+//
+//     // return interface->getCcd().getPredictiveLogLikelihood(&weights[0]);
+//     Rcpp::stop("No longer implemented");
+//     return 0.0;
+// }
 
 // [[Rcpp::export(".cyclopsGetNewPredictiveLogLikelihood")]]
 double cyclopsGetNewPredictiveLogLikelihood(SEXP inRcppCcdInterface,
@@ -193,6 +197,15 @@ double cyclopsGetLogLikelihood(SEXP inRcppCcdInterface) {
 	XPtr<RcppCcdInterface> interface(inRcppCcdInterface);
 
 	return interface->getCcd().getLogLikelihood();
+}
+
+
+// [[Rcpp::export(".cyclopsLogResults")]]
+void cyclopsLogResult(SEXP inRcppCcdInterface, const std::string& fileName, bool withASE) {
+    using namespace bsccs;
+    XPtr<RcppCcdInterface> interface(inRcppCcdInterface);
+
+    interface->logResultsToFile(fileName, withASE);
 }
 
 // [[Rcpp::export(".cyclopsGetFisherInformation")]]
@@ -631,8 +644,8 @@ void RcppCcdInterface::appendRList(Rcpp::List& list, const Rcpp::List& append) {
 }
 
 void RcppCcdInterface::handleError(const std::string& str) {
-//	Rcpp::stop(str); // TODO Want this to work
-	::Rf_error(str.c_str());
+	Rcpp::stop(str);
+//	::Rf_error(str.c_str());
 }
 
 bsccs::ConvergenceType RcppCcdInterface::parseConvergenceType(const std::string& convergenceName) {
@@ -679,6 +692,8 @@ bsccs::priors::PriorType RcppCcdInterface::parsePriorType(const std::string& pri
 		priorType = NORMAL;
 	} else if (priorName == "barupdate") {
 	    priorType = BAR_UPDATE;
+	} else if (priorName == "jeffreys") {
+	    priorType = JEFFREYS;
 	} else {
  		handleError("Invalid prior type.");
  	}
@@ -953,7 +968,7 @@ void RcppCcdInterface::initializeModelImpl(
 //  	}
 //  	singlePrior->setVariance(0, arguments.hyperprior);
 
- 	JointPriorPtr prior;
+ 	JointPriorPtr prior = nullptr;
 //  	if (arguments.flatPrior.size() == 0) {
 //  		prior = bsccs::make_shared<FullyExchangeableJointPrior>(singlePrior);
 //  	} else {