+++ src/cyclops/drivers/AutoSearchCrossValidationDriver.cpp
@@ -112,10 +112,13 @@ MaxPoint AutoSearchCrossValidationDriver::doCrossValidationLoop(
 
 	            std::vector<double> predLogLikelihood;
 
+	            std::cerr << "A" << std::endl;
+
 	            // Newly re-located code
 	            double pointEstimate = doCrossValidationStep(ccd, selector, allArguments, step,
                                                           nThreads, ccdPool, selectorPool,
                                                           predLogLikelihood);
+                std::cerr << "B" << std::endl;
 
 	            double stdDevEstimate = computeStDev(predLogLikelihood, pointEstimate);