+++ tests/testthat/test-smallFineGray.R
@@ -5,6 +5,7 @@ library("cmprsk")
 suppressWarnings(RNGversion("3.5.0"))
 
 test_that("Check that Cox fits only have 2 outcome identifies", {
+    skip_on_cran()
     test <- read.table(header=T, sep = ",", text = "
                        start, length, event, x1, x2
                        0, 4,  1,0,0
@@ -123,6 +124,7 @@ test_that("Check very small Fine-Gray example with no ties (sparse vs. dense)", 
 })
 
 test_that("Check very small Fine-Gray example with no ties using Andromeda", {
+    skip_on_cran()
     test <- read.table(header=T, sep = ",", text = "
                        start, length, event, x1, x2
                        0, 4,  1,0,0
@@ -252,7 +254,7 @@ test_that("Check very small Fine-Gray versus Cox for 0/1 events (censorWeights d
     dataPtr <- Cyclops:::createCyclopsData(fgDat$surv ~ test$x1 + test$x2, modelType = "fgr", censorWeights = fgDat$weights)
     cyclopsFit1 <- Cyclops:::fitCyclopsModel(dataPtr)
     dataPtr <- Cyclops:::createCyclopsData(fgDat$surv ~ test$x1 + test$x2, modelType = "cox", censorWeights = fgDat$weights)
-    cyclopsFit2 <- Cyclops:::fitCyclopsModel(dataPtr)
+    cyclopsFit2 <- Cyclops:::fitCyclopsModel(dataPtr, warnings = FALSE)
 
     tolerance <- 1E-4
     expect_equivalent(coef(cyclopsFit1), coef(cyclopsFit2), tolerance = tolerance)
@@ -272,7 +274,7 @@ test_that("Check very small Cox example with weights and censor weights defined 
     fgDat <- Cyclops:::getFineGrayWeights(test$length, test$event)
     dataPtr <- Cyclops:::createCyclopsData(fgDat$surv ~ test$x1 + test$x2, modelType = "cox",
                                            weights = c(1, 1, 1, 0, 0, 1, 1), censorWeights = fgDat$weights)
-    cyclopsFit1 <- Cyclops:::fitCyclopsModel(dataPtr)
+    cyclopsFit1 <- Cyclops:::fitCyclopsModel(dataPtr, warnings = FALSE)
     dataPtr <- Cyclops:::createCyclopsData(fgDat$surv ~ test$x1 + test$x2, modelType = "cox",
                                            weights = c(1, 1, 1, 0, 0, 1, 1))
     cyclopsFit2 <- Cyclops:::fitCyclopsModel(dataPtr)